Add how-to for installing a zone cluster
This how-to mostly refers to the OpenShift 4 installation how-tos on
kb.vshn.ch/oc4/ but documents the customizations that are required when
setting up a cluster as an APPUiO Cloud zone.

Mainly the customizations ensure that the cluster uses the APPUiO Cloud
DNS scheme instead of the APPUiO Managed OCP4 DNS scheme.