Remove `org-` prefix for organization namespaces
There does not seem to be consensus on what naming collisions this prefix
should prevent.

The prefix makes the API less intuitive. When working with organization
`foo`, I expect to be able to access its teams in namespace `foo` and
not namespace `org-foo`. It's even worse when there actually is a
namespace `foo` and I get a permission denied error.